Moved all .json files to /configs/
Added event to have the bot watch what room she is in and if she is alone to take action.  Said actions are no ready yet.

Reworked the !play logic to append to listSubmitted and watched the logic on user submitted tracks

Ran the code for two hours on its basic loop with no errors.